+++ flash/ek/flash/rasterizer/cairo_utility.hpp
@@ -47,9 +47,9 @@ cairo_pattern_t* create_radial_pattern(const matrix_2d& matrix);
 
 fill_pattern_data_t create_fill_pattern(const fill_style& fill, const transform_model& transform);
 
-void cairo_round_rectangle(cairo_t* cr, const float* values);
+void cairo_round_rectangle(cairo_t* cr, const double* values);
 
-void cairo_oval(cairo_t* cr, const float* values);
+void cairo_oval(cairo_t* cr, const double* values);
 
 void cairo_transform(cairo_t* cr, const matrix_2d& m);