+++ scenex/ek/scenex/data/sg_data.cpp
@@ -5,9 +5,9 @@
 namespace ek {
 
 const sg_node_data* sg_file::get(const std::string& library_name) const {
-    for (auto& child : library.children) {
-        if (child.libraryName == library_name) {
-            return &child;
+    for (auto& item : library) {
+        if (item.libraryName == library_name) {
+            return &item;
         }
     }
     return nullptr;