+++ flash/CHANGELOG.md
@@ -15,3 +15,7 @@
 - Add stroke gradients
 - Fix gradient matrix transform
 - Add Bitmap fills render
+- Change fill bitmap reference resolving to parsing phase
+- Add `doc_parser` to preserve `doc` instance for parsing methods
+- Change render command values to `double`
+- Add `delete` of `export_item_t` instances